+++ bc/recruitment_api/management/commands/import_jobs.py
@@ -1,9 +1,11 @@
+from django.core.files.base import ContentFile
 from django.core.management.base import BaseCommand
 from django.utils.timezone import now
 
+from bc.documents.models import CustomDocument
 from bc.recruitment.models import TalentLinkJob
 from bc.recruitment_api.client import get_client
-from bc.recruitment_api.utils import update_job_from_ad
+from bc.recruitment_api.utils import delete_jobs, update_job_from_ad
 
 
 class Command(BaseCommand):
@@ -23,7 +25,9 @@ class Command(BaseCommand):
         results = True
         num_updated = 0
         num_created = 0
+        doc_imported = 0
         errors = []
+        import_timestamp = now()
         while results:
             self.stdout.write(f"Fetching page {page}")
             response = client.service.getAdvertisementsByPage(
@@ -46,7 +50,7 @@ class Command(BaseCommand):
                         job = update_job_from_ad(
                             job,
                             ad,
-                            defaults={"last_imported": now()},
+                            defaults={"last_imported": import_timestamp},
                             import_categories=options["import_categories"],
                         )
                     except Exception as e:
@@ -61,9 +65,52 @@ class Command(BaseCommand):
                         else:
                             num_updated += 1
 
+                        # Fetch attachments via a different call
+                        try:
+                            # This will return list of attachments with
+                            #   'content', 'description', 'fileName', 'id', 'mimeType'
+                            attachments_response = client.service.getAttachments(
+                                job.talentlink_id
+                            )
+                            for attachment in attachments_response:
+                                if attachment["id"] and attachment["fileName"]:
+                                    doc, created = CustomDocument.objects.get_or_create(
+                                        talentlink_attachment_id=attachment["id"]
+                                    )
+                                    if created:
+                                        doc.title = (
+                                            attachment["description"]
+                                            or attachment["fileName"].split(".")[0]
+                                        )
+                                        doc.file = ContentFile(
+                                            attachment["content"],
+                                            name=attachment["fileName"],
+                                        )
+                                        doc.save()
+                                        doc_imported += 1
+
+                                    job.attachments.add(doc)
+                                    job.save()
+                        except Exception as e:
+                            msg = (
+                                f"Error occurred while importing attachments for job {ad['id']}:\n"
+                                + str(e)
+                            )
+                            errors.append(msg)
+
+        # Check for outdated jobs
+        num_deleted = 0
+        try:
+            num_deleted = delete_jobs(import_timestamp)
+        except Exception as e:
+            msg = f"Error occurred while deleting jobs:\n" + str(e)
+            errors.append(msg)
+
         self.stdout.write("No more results")
         self.stdout.write(f"{num_updated} existing jobs updated")
         self.stdout.write(f"{num_created} new jobs created")
+        self.stdout.write(f"{doc_imported} new documents imported")
+        self.stdout.write(f"{num_deleted} jobs deleted")
         self.stdout.write(self.style.ERROR(f"{len(errors)} errors"))
         for error in errors:
             self.stdout.write(self.style.ERROR(msg))