+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Release History
 
-## Unreleased
+## 10.0 (2020-03-31)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/8.0...9.0
+
+- 114 StreamField Button Link block (BE)
+- 125 Job attachments
+- 138 Additional job sidebar fields (FE)
+- 139 Additional job sidebar fields (BE)
+- 148 Delete outdated jobs when running import command
 
 ## 9.0 (2020-03-31)